[DPTOOLS-66] Read 1.8 whitelist from txt file in etl (#24)
Modifying the 1.8 whitelist currently requires changes in 3 places:

1. .airflowignore in the `etl` repo
2. jobs.py in `incubator-airflow/data-platform`
3. jobs.py in `incubator-airflow/data-platform-upgrade`

This reads this configuration from a config file in the `etl` repo so that
we only need to modify the `etl` repo to update the whitelist. Once this
is approved, I'll submit a PR in etl to pin to the new Airflow SHA and
add the whitelist .txt file.

Then I'll port this exact same logic to our airflow 1.8 branch.