[DPTOOLS-82] Restore REMOVED TI's if present in DAG
This PR is intended to address https://issues.apache.org/jira/browse/AIRFLOW-1460

If a TI is present in the database but it is *not* present in the DAG
(in-memory), the scheduler marks the task as REMOVED. Once a TI was
removed, it stayed removed forever, even if the TI was added back to the
DAG later.

This caused us problems with DAGS built dynamically based on the
contents of the `etl_tables` table in `lyftpg`. If any one of the 1000's
of calls to postgres fails during the ~12 hours while `other` DAG is
running, all of the TIs in the running DAG are marked as REMOVED. This
failure scenario applies to any DAG which relies on an external data
store to self-populate.

With this change, we will allow for the restoration of TI's in the
REMOVED state. If a TI is REMOVED in the database but it is now present
in the DAG, we restore the TI by setting the state to NONE.

= Testing =

Added a new unit test, which you can run in a venv like this:

./run_unit_tests.sh tests.models:DagRunTest -s