[DPTOOLS-82] Ignore missing TI if already removed
Our task removal metrics [1] are skewed because we are constantly
marking the same tasks as removed over and over again. If we see that a
TI that is not in the DAG is already in state `REMOVED` in the database,
we can simply ignore it without increasing our counter or logging an
exception.

[1] https://lyft.wavefront.com/u/chvzlGtKQh