[AIRFLOW-2520] CLI - make backfill less verbose
Cherry-picking this commit [1] from upstream.

[1] https://github.com/apache/incubator-airflow/commit/972086aeba4616843005b25210ba3b2596963d57

Original commit message:

Used backfill recently and it would log a shit ton of logging messages
telling me all the tasks that were not ready to run at every tick.

These messages are not useful and should be muted by default.

I understand that this may be helpful in the context of `airflow run` in
the context where dependencies aren't met, so decided to manage a flag
instead of simply going `logging.debug` on it.

Closes #3414 from mistercrunch/backfill_less_verbose