Fix bug when setting task instance state
Pull in this change [1] from upstream. Original commit message follows:

When there is a dot in a task's ID or in its execution date, changing
its state from the UI fails.

The reason for this is that Flask-Admin's iterencode function encodes
composite primary keys using dot as the escape symbol. If there is
already a dot in any of the values, it is doubled.

On decoding however, Airflow is currently just splitting the string on a
comma.

This PR switches to using Flask-Admin's iterdecode to correctly decode
task instance rowids when they contain a dot.

[1] https://github.com/apache/incubator-airflow/commit/a27bd620d97950773c7732908eb3c597315fdd0e