Allow Public role access to /home
Problem: After a user registers in the PUBLIC_AUTH_ROLE they are
redirected to /home. But a user in the Public role still doesn't have
view privileges on /home, so they are immediately
redirected *back* to the /login page, which creates an endless redirect
cycle.

This PR change grants the Public role view access to the index page.
Because the Public user doesn't have any other read permission, they
won't see any DAGs.

If the user is not logged in, we add a banner at the top of the page
with a link to the login page. If the user is logged in but belongs only
to the Public role, then the banner displays a message warning the user
that they don't have sufficient privileges to view the DAGs.